Logout link and proper login redirects
