+++ apps/dockup_ui/web/controllers/auth_controller.ex
@@ -13,13 +13,13 @@ defmodule DockupUi.AuthController do
     conn
     |> put_flash(:info, "You have been logged out!")
     |> configure_session(drop: true)
-    |> redirect(to: "/")
+    |> redirect(to: unauthenticated_path(conn))
   end
 
   def callback(%{assigns: %{ueberauth_failure: _fails}} = conn, _params) do
     conn
     |> put_flash(:error, "Failed to authenticate.")
-    |> redirect(to: "/")
+    |> redirect(to: unauthenticated_path(conn))
   end
 
   def callback(%{assigns: %{ueberauth_auth: auth}} = conn, _params) do
@@ -28,11 +28,15 @@ defmodule DockupUi.AuthController do
         conn
         |> put_flash(:info, "Successfully authenticated.")
         |> put_session(:current_user, user)
-        |> redirect(to: "/")
+        |> redirect(to: deployment_path(conn, :new))
       {:error, reason} ->
         conn
         |> put_flash(:error, reason)
-        |> redirect(to: "/")
+        |> redirect(to: unauthenticated_path(conn))
     end
+  end
+
+  defp unauthenticated_path(conn) do
+    deployment_path(conn, :home)
   end
 end
