+++ apps/dockup_ui/web/templates/layout/app.html.eex
@@ -19,6 +19,7 @@
         <div class="navbar-nav">
           <%= link("Recent Deployments", to: deployment_path(@conn, :index), class: "nav-item nav-link") %>
           <%= link("Configure", to: config_path(@conn, :index), class: "nav-item nav-link")%>
+          <%= link("Logout", to: auth_path(@conn, :delete), method: :delete, class: "nav-item nav-link") %>
         </div>
       </div>
     </nav>
